Set up digital
