+++ gui/consts.py
@@ -1,0 +1,5 @@
+from logging import getLogger
+import re
+
+LOGGER = getLogger("inodaqv2")
+PAT_VALID_DIG = re.compile(r"^1;\d{1,2},(on|off)$")
